pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "setuptools-scm>=7.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deeplightrag"
version = "1.0.14"
description = "DeepLightRAG: High-performance Document Indexing and Retrieval System (use with any LLM)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Phuong Nguyen", email = "nhphuong.code@gmail.com"}
]
maintainers = [
    {name = "Phuong Nguyen", email = "nhphuong.code@gmail.com"}
]
keywords = [
    "rag",
    "retrieval",
    "augmented",
    "generation",
    "ocr",
    "vision",
    "graph",
    "nlp",
    "llm",
    "deepseek",
    "document-processing"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Text Processing",
    "Typing :: Typed",
]

dependencies = [
    # Core dependencies
    "numpy>=1.24.0",
    "networkx>=3.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "tqdm>=4.65.0",
    "typing-extensions>=4.0.0;python_version<'3.10'",
    
    # PDF Processing
    "pdf2image>=1.16.0",
    "PyMuPDF>=1.23.0",
    "easyocr>=1.7.0",
    
    # ML/AI Core
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0,<4.47.0",
    "accelerate>=0.24.0",
    "sentence-transformers>=2.2.0",
    "setfit>=1.0.0",
    
    # NER and Relation Extraction
    "gliner>=0.1.12",
    
    # Vector Storage
    "faiss-cpu>=1.7.4",
    "toon-python>=0.1.2",
]

[project.optional-dependencies]
# GPU acceleration (CUDA support)
gpu = [
    "bitsandbytes>=0.41.0",
]

# MLX support (macOS Apple Silicon only)
macos = [
    "mlx>=0.21.0",
    "mlx-lm>=0.19.0",
    "mlx-vlm>=0.0.3",
]

# LLM providers (optional - use if you want built-in generation)
llm = [
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
]

# Advanced relation extraction
advanced-re = [
    "opennre>=1.1.0",
]

# Web interface
web = [
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "pandas>=2.0.0",
]

# Development tools
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-timeout>=2.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.9.0",
    "ruff>=0.0.290",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "wheel>=0.41.0",
]

# Documentation
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=1.0.0",
]

# All optional dependencies
all = [
    "deeplightrag[gpu,llm,advanced-re,web,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/png261/DeepLightRag"
Repository = "https://github.com/png261/DeepLightRag"
"Bug Tracker" = "https://github.com/png261/DeepLightRag/issues"
Changelog = "https://github.com/png261/DeepLightRag/releases"

[project.scripts]
deeplightrag = "deeplightrag.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["deeplightrag*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
deeplightrag = ["py.typed", "*.yaml", "*.json"]

# Black code formatter configuration
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312", "py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

# Ruff linter configuration
[tool.ruff]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line-too-long (handled by black)
    "B008",   # function-call-in-default-argument
    "C901",   # complex-structure
]
line-length = 100
target-version = "py39"
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "build",
    "dist",
    "migrations",
    "__pycache__",
]

[tool.ruff.isort]
known-first-party = ["deeplightrag"]
force-single-line = false

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/*" = ["F841", "F401"]

# MyPy type checking configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
exclude = [
    "tests/",
    "examples/",
    "build/",
    "dist/",
    ".venv/",
]

[[tool.mypy.overrides]]
module = [
    "PIL.*",
    "pdf2image.*",
    "fitz.*",
    "docx.*",
    "openpyxl.*",
    "pptx.*",
    "markdown.*",
    "bs4.*",
    "langchain.*",
    "mlx.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=src/deeplightrag",
    "--cov-report=xml",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=70",
    "-W ignore::DeprecationWarning",
]
minversion = "7.0"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
norecursedirs = [".git", ".tox", "dist", "build", "*.egg"]

# Coverage configuration
[tool.coverage.run]
branch = true
source = ["src/deeplightrag"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/cli.py",
    "*/main.py",
]
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
    "if sys.version_info",
    "if platform.system",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

# Pylint configuration
[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "W0212",  # protected-access
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
    "W0511",  # fixme
    "R0801",  # duplicate-code
    "C0302",  # too-many-lines
]

[tool.pylint.format]
max-line-length = 100
max-module-lines = 2000

[tool.pylint.basic]
good-names = ["i", "j", "k", "ex", "Run", "_", "x", "y", "z", "df", "f", "ax", "e"]

# Bandit security configuration
[tool.bandit]
exclude_dirs = ["tests", "examples", ".venv"]
skips = ["B101"]  # assert_used

# Interrogate docstring coverage
[tool.interrogate]
ignore-init-method = false
ignore-init-module = false
ignore-magic = true
ignore-semiprivate = false
ignore-private = false
fail-under = 50
verbose = 1
quiet = false
whitelist-regex = "test_.*"
